requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
opencv-python
pillow
torch
torchvision
tqdm
requests
